+++ CHANGELOG.md
@@ -1,3 +1,15 @@
+## 5.2.2-customercenter.alpha.3
+
+- Fix for disabled promo offer button (#4142) 
+
+## 5.2.2-customercenter.alpha.2
+
+- Fix project.pbxproj (#4122)
+- Fix BackendGetCustomerCenterConfigTests (#4124)
+- Add contact support button (#4023) 
+- Fix checking eligibility (#4138)
+- Make colors nullable (#4134)
+
 ## 5.2.3
 ### Bugfixes
 * Fix Paywalls crash on iOS 18 beta (#4154) via Andy Boedo (@aboedo)
@@ -29,6 +41,10 @@
 - Fixes a typo in the bug_report issue template (#3945) via JayShortway (@JayShortway)
 - [External] Add missing SwiftUI environment for previews (#4109) via @noahsmartin (#4110) via Andy Boedo (@aboedo)
 - Remove notify-on-non-patch-release-branches (#4106) via Cesar de la Vega (@vegaro)
+
+## 5.2.1-customercenter.alpha.1
+
+- Initial Customer Center Alpha Release
 
 ## 5.2.1
 ### Bugfixes